zlib: build separate shared/static copies
When both shared and static libraries were built, the library paths were
recorded in ZLIB.cmake and, when the static library was split out, it
causes all sources depending on zlib-ng to fail due to CMake errors.

As a workaround, build full shared and static copies. For the static copy,
only install libz.a and nothing else. This way, we get a ZLIB.cmake
consistent with what is installed in the main package with shared
libraries, and a libz.a used by packages that need it.