postgresql-17: new, 17.5; Transition from postgresql
We are now shipping PostgreSQL with multiple versions. This involves a
packageing strategy change, where the `postgresql' package itself no
longer holds the server it self, and instead is a metapackage that
points to the current latest major release, e.g. `postgresql-17'.

Also we are splitting PostgreSQL into runtime and server packages, so
there's no unnecessary things installed to the system (Qt depends on
PostgreSQL runtime libraries, but most of the time users don't need to
run the PostgreSQL server itself.)

* Split the server into runtime and server programs.
* Remove unused logrotate.d and PAM configuration.
* Deploy systemd unit files with versioned suffixes.
* Mark PKGBREAK and PKGREP.