linux-kernel: Drop the LoongArch SMCv0 CPUFreq patch
It causes "cpufreq: __target_index: Failed to change cpu frequency: -1"
on some 2K3000 and 3A6000 systems, including one 3A6000 which hangs
after spitting a lot of this error.

And this patch was not introduced properly in the first place: there
are two "0334" in the patchset.  We should really make autobuild error
out on this.

To me this it seems this patch has just accidentally entered the patch
set, but then it was reported in a Bi-weekly.  I'm really puzzled but
let's kick it out (at least for now) anyway.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>