aosc-os-presets-desktop: Don't enable avahi-daemon.service by default
systemd-resolved already has the mDNS transponder functionality, and
it's enabled by default.  Thus having systemd-resolved running together
with avahi-daemon will cause a conflict.

The conflict is a contribute factor of the very annoying "avahi-daemon
keeps reporting a conflict and bump the ID in the host name" issue, but
that issue is not solely caused by the conflict (i.e. it even happens
without the conflict).  Considering this issue is still not fixed (nor
even understood) after years it was reported, to me we should just keep
using the systemd mDNS transponder and disable avahi by default.

If a user must use avahi they need to at least set MulticastDNS=no in
systemd/resolved.conf.d/ and if they still hit the annoying issue I
guess we've no way to help them.